Display profile tab only when logged in
